Extract business logic for balance screen into viewmodel
